# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Global Football Capital — multi-player football economy platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register user (pick role and display name)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Leagues ───────────────────────────────────────────────────────────────
...
  /clubs:
    get:
      operationId: listClubs
      tags: [clubs]
      summary: Browse all clubs
      parameters:
        - name: leagueId
          in: query
          schema:
            type: integer
        - name: available
          in: query
          description: Only clubs without an owner
          schema:
            type: boolean
      responses:
        "200":
          description: Clubs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Club"
  /clubs/{id}:
    get:
      operationId: getClub
      tags: [clubs]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Club detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClubDetail"
  /clubs/{id}/buy:
    post:
      operationId: buyClub
      tags: [clubs]
      summary: Owner purchases a club (costs GFC coins)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Club acquired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClubDetail"
        "400":
          description: Insufficient funds or already owned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /clubs/{id}/hire-coach:
    post:
      operationId: hireCoach
      tags: [clubs]
      summary: Owner hires a coach for their club
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HireCoachInput"
      responses:
        "200":
          description: Coach hired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClubDetail"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /clubs/{id}/fire-coach:
    post:
      operationId: fireCoach
      tags: [clubs]
      summary: Owner fires the current coach (penalty applies)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Coach fired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClubDetail"
  /clubs/{id}/upgrade:
    post:
      operationId: upgradeClub
      tags: [clubs]
      summary: Owner upgrades infrastructure (stadium, academy, training)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpgradeInput"
      responses:
        "200":
          description: Upgraded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClubDetail"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /clubs/my:
    get:
      operationId: getMyClub
      tags: [clubs]
      summary: Get the club the current user owns or coaches
      responses:
        "200":
          description: My club
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClubDetail"
        "404":
          description: No club
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Players ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
        clerkId:
          type: string
        displayName:
          type: string
        role:
          type: string
          enum: [owner, coach, unregistered]
        balance:
          type: integer
          description: GFC coins
        reputation:
          type: integer
        clubId:
          type: ["integer", "null"]
        clubName:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, clerkId, displayName, role, balance, reputation, createdAt]
    RegisterInput:
      type: object
      properties:
        displayName:
          type: string
        role:
          type: string
          enum: [owner, coach]
      required: [displayName, role]

    # ── Leagues ─────────────────────────────────────────────────────────────
...
    Club:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        leagueId:
          type: integer
        leagueName:
          type: string
        country:
          type: string
        city:
          type: string
        colors:
          type: string
        purchasePrice:
          type: integer
        currentValue:
          type: integer
        prestige:
          type: integer
        ownerId:
          type: ["string", "null"]
        ownerName:
          type: ["string", "null"]
        coachId:
          type: ["string", "null"]
        coachName:
          type: ["string", "null"]
        stadiumLevel:
          type: integer
        academyLevel:
          type: integer
        trainingLevel:
          type: integer
        transferBudget:
          type: integer
        weeklyWageBill:
          type: integer
      required: [id, name, leagueId, leagueName, country, city, colors, purchasePrice, currentValue, prestige, stadiumLevel, academyLevel, trainingLevel, transferBudget, weeklyWageBill]
    ClubDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        leagueId:
          type: integer
        leagueName:
          type: string
        country:
          type: string
        city:
          type: string
        colors:
          type: string
        purchasePrice:
          type: integer
        currentValue:
          type: integer
        prestige:
          type: integer
        ownerId:
          type: ["string", "null"]
        ownerName:
          type: ["string", "null"]
        coachId:
          type: ["string", "null"]
        coachName:
          type: ["string", "null"]
        stadiumLevel:
          type: integer
        academyLevel:
          type: integer
        trainingLevel:
          type: integer
        transferBudget:
          type: integer
        weeklyWageBill:
          type: integer
        stadiumName:
          type: string
        stadiumCapacity:
          type: integer
        description:
          type: string
        players:
          type: array
          items:
            $ref: "#/components/schemas/Player"
        coachSalary:
          type: ["integer", "null"]
        coachContractWeeks:
          type: ["integer", "null"]
      required: [id, name, leagueId, leagueName, country, city, colors, purchasePrice, currentValue, prestige, stadiumLevel, academyLevel, trainingLevel, transferBudget, weeklyWageBill, stadiumName, stadiumCapacity, description, players]
    UpgradeInput:
      type: object
      properties:
        upgradeType:
          type: string
          enum: [stadium, academy, training]
      required: [upgradeType]
    HireCoachInput:
      type: object
      properties:
        coachUserId:
          type: string
        weeklySalary:
          type: integer
        contractWeeks:
          type: integer
        bonusPerWin:
          type: integer
      required: [coachUserId, weeklySalary, contractWeeks, bonusPerWin]

    # ── Players ─────────────────────────────────────────────────────────────
    Player:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        age:
          type: integer
        nationality:
          type: string
        position:
          type: string
        overall:
          type: integer
        potential:
          type: integer
        pace:
          type: integer
        shooting:
          type: integer
        passing:
          type: integer
        dribbling:
          type: integer
        defending:
          type: integer
        physicality:
          type: integer
        value:
          type: integer
        weeklySalary:
          type: integer
        contractEndsWeek:
          type: ["integer", "null"]
        clubId:
          type: ["integer", "null"]
        clubName:
          type: ["string", "null"]
        isOnTransferList:
          type: boolean
        isFreeAgent:
          type: boolean
      required: [id, name, age, nationality, position, overall, potential, pace, shooting, passing, dribbling, defending, physicality, value, weeklySalary, isOnTransferList, isFreeAgent]

    # ── Contracts ───────────────────────────────────────────────────────────
...